feat: allow to run release.yml on workflow dispatch
If something went wrong you now have a second chance.
And you can release from a release branch to apply
bugfixes to old releases.

You need to cherry pick this commit
onto the release branch that this is possible on old branches.
This ports the fix of:
https://github.com/appuio/charts/pull/535
To this repo.